+++ opentelemetry-otlp/examples/basic-otlp-http/src/main.rs
@@ -11,7 +11,7 @@ use opentelemetry_otlp::{LogExporter, MetricExporter, Protocol, SpanExporter};
 use opentelemetry_sdk::{
     logs::SdkLoggerProvider,
     metrics::{MetricError, SdkMeterProvider},
-    trace::{self as sdktrace, TracerProvider},
+    trace::{self as sdktrace, SdkTracerProvider},
 };
 use opentelemetry_sdk::{
     logs::{self as sdklogs},
@@ -41,14 +41,14 @@ fn init_logs() -> Result<sdklogs::SdkLoggerProvider, opentelemetry_sdk::logs::Lo
         .build())
 }
 
-fn init_traces() -> Result<sdktrace::TracerProvider, TraceError> {
+fn init_traces() -> Result<sdktrace::SdkTracerProvider, TraceError> {
     let exporter = SpanExporter::builder()
         .with_http()
         .with_protocol(Protocol::HttpBinary) //can be changed to `Protocol::HttpJson` to export in JSON format
         .with_endpoint("http://localhost:4318/v1/traces")
         .build()?;
 
-    Ok(TracerProvider::builder()
+    Ok(SdkTracerProvider::builder()
         .with_batch_exporter(exporter)
         .with_resource(RESOURCE.clone())
         .build())