+++ opentelemetry-sdk/benches/span_builder.rs
@@ -53,8 +53,8 @@ fn span_builder_benchmark_group(c: &mut Criterion) {
     group.finish();
 }
 
-fn not_sampled_provider() -> (sdktrace::TracerProvider, sdktrace::Tracer) {
-    let provider = sdktrace::TracerProvider::builder()
+fn not_sampled_provider() -> (sdktrace::SdkTracerProvider, sdktrace::SdkTracer) {
+    let provider = sdktrace::SdkTracerProvider::builder()
         .with_sampler(sdktrace::Sampler::AlwaysOff)
         .with_simple_exporter(NoopExporter)
         .build();