+++ opentelemetry-sdk/src/trace/provider.rs
@@ -36,7 +36,7 @@ use crate::error::{OTelSdkError, OTelSdkResult};
 ///
 /// ```
 /// use opentelemetry::global;
-/// use opentelemetry_sdk::trace::TracerProvider;
+/// use opentelemetry_sdk::trace::SdkTracerProvider;
 /// use opentelemetry::trace::Tracer;
 ///
 /// fn init_tracing() -> TracerProvider {
@@ -64,7 +64,8 @@ use crate::error::{OTelSdkError, OTelSdkResult};
 /// }
 /// ```
 use crate::trace::{
-    BatchSpanProcessor, Config, RandomIdGenerator, Sampler, SimpleSpanProcessor, SpanLimits, Tracer,
+    BatchSpanProcessor, Config, RandomIdGenerator, Sampler, SdkTracer, SimpleSpanProcessor,
+    SpanLimits,
 };
 use crate::Resource;
 use crate::{trace::SpanExporter, trace::SpanProcessor};
@@ -80,11 +81,11 @@ static PROVIDER_RESOURCE: OnceLock<Resource> = OnceLock::new();
 
 // a no nop tracer provider used as placeholder when the provider is shutdown
 // TODO Replace with LazyLock once it is stable
-static NOOP_TRACER_PROVIDER: OnceLock<TracerProvider> = OnceLock::new();
+static NOOP_TRACER_PROVIDER: OnceLock<SdkTracerProvider> = OnceLock::new();
 #[inline]
-fn noop_tracer_provider() -> &'static TracerProvider {
+fn noop_tracer_provider() -> &'static SdkTracerProvider {
     NOOP_TRACER_PROVIDER.get_or_init(|| {
-        TracerProvider {
+        SdkTracerProvider {
             inner: Arc::new(TracerProviderInner {
                 processors: Vec::new(),
                 config: Config {
@@ -142,7 +143,7 @@ impl Drop for TracerProviderInner {
     }
 }
 
-/// Creator and registry of named [`Tracer`] instances.
+/// Creator and registry of named [`SdkTracer`] instances.
 ///
 /// `TracerProvider` is a container holding pointers to `SpanProcessor` and other components.
 /// Cloning a `TracerProvider` instance and dropping it will not stop span processing. To stop span processing, users
@@ -150,25 +151,25 @@ impl Drop for TracerProviderInner {
 /// to be dropped. When the last reference is dropped, the shutdown process will be automatically triggered
 /// to ensure proper cleanup.
 #[derive(Clone, Debug)]
-pub struct TracerProvider {
+pub struct SdkTracerProvider {
     inner: Arc<TracerProviderInner>,
 }
 
-impl Default for TracerProvider {
+impl Default for SdkTracerProvider {
     fn default() -> Self {
-        TracerProvider::builder().build()
+        SdkTracerProvider::builder().build()
     }
 }
 
-impl TracerProvider {
+impl SdkTracerProvider {
     /// Build a new tracer provider
     pub(crate) fn new(inner: TracerProviderInner) -> Self {
-        TracerProvider {
+        SdkTracerProvider {
             inner: Arc::new(inner),
         }
     }
 
-    /// Create a new [`TracerProvider`] builder.
+    /// Create a new [`SdkTracerProvider`] builder.
     pub fn builder() -> TracerProviderBuilder {
         TracerProviderBuilder::default()
     }
@@ -195,7 +196,7 @@ impl TracerProvider {
     ///
     /// ```
     /// use opentelemetry::global;
-    /// use opentelemetry_sdk::trace::TracerProvider;
+    /// use opentelemetry_sdk::trace::SdkTracerProvider;
     ///
     /// fn init_tracing() -> TracerProvider {
     ///     let provider = TracerProvider::default();
@@ -264,9 +265,9 @@ impl TracerProvider {
     }
 }
 
-impl opentelemetry::trace::TracerProvider for TracerProvider {
+impl opentelemetry::trace::TracerProvider for SdkTracerProvider {
     /// This implementation of `TracerProvider` produces `Tracer` instances.
-    type Tracer = Tracer;
+    type Tracer = SdkTracer;
 
     fn tracer(&self, name: impl Into<Cow<'static, str>>) -> Self::Tracer {
         let scope = InstrumentationScope::builder(name).build();
@@ -275,12 +276,12 @@ impl opentelemetry::trace::TracerProvider for TracerProvider {
 
     fn tracer_with_scope(&self, scope: InstrumentationScope) -> Self::Tracer {
         if self.inner.is_shutdown.load(Ordering::Relaxed) {
-            return Tracer::new(scope, noop_tracer_provider().clone());
+            return SdkTracer::new(scope, noop_tracer_provider().clone());
         }
         if scope.name().is_empty() {
             otel_info!(name: "TracerNameEmpty",  message = "Tracer name is empty; consider providing a meaningful name. Tracer will function normally and the provided name will be used as-is.");
         };
-        Tracer::new(scope, self.clone())
+        SdkTracer::new(scope, self.clone())
     }
 }
 
@@ -401,10 +402,10 @@ impl TracerProviderBuilder {
         self
     }
 
-    /// Associates a [Resource] with a [TracerProvider].
+    /// Associates a [Resource] with a [SdkTracerProvider].
     ///
     /// This [Resource] represents the entity producing telemetry and is associated
-    /// with all [Tracer]s the [TracerProvider] will create.
+    /// with all [Tracer]s the [SdkTracerProvider] will create.
     ///
     /// By default, if this option is not used, the default [Resource] will be used.
     ///
@@ -417,7 +418,7 @@ impl TracerProviderBuilder {
     }
 
     /// Create a new provider from this configuration.
-    pub fn build(self) -> TracerProvider {
+    pub fn build(self) -> SdkTracerProvider {
         let mut config = self.config;
 
         // Standard config will contain an owned [`Resource`] (either sdk default or use supplied)
@@ -445,7 +446,7 @@ impl TracerProviderBuilder {
         }
 
         let is_shutdown = AtomicBool::new(false);
-        TracerProvider::new(TracerProviderInner {
+        SdkTracerProvider::new(TracerProviderInner {
             processors,
             config,
             is_shutdown,
@@ -543,7 +544,7 @@ mod tests {
 
     #[test]
     fn test_force_flush() {
-        let tracer_provider = super::TracerProvider::new(TracerProviderInner {
+        let tracer_provider = super::SdkTracerProvider::new(TracerProviderInner {
             processors: vec![
                 Box::from(TestSpanProcessor::new(true)),
                 Box::from(TestSpanProcessor::new(false)),
@@ -558,7 +559,7 @@ mod tests {
 
     #[test]
     fn test_tracer_provider_default_resource() {
-        let assert_resource = |provider: &super::TracerProvider,
+        let assert_resource = |provider: &super::SdkTracerProvider,
                                resource_key: &'static str,
                                expect: Option<&'static str>| {
             assert_eq!(
@@ -570,7 +571,7 @@ mod tests {
                 expect.map(|s| s.to_string())
             );
         };
-        let assert_telemetry_resource = |provider: &super::TracerProvider| {
+        let assert_telemetry_resource = |provider: &super::SdkTracerProvider| {
             assert_eq!(
                 provider
                     .config()
@@ -593,7 +594,7 @@ mod tests {
 
         // If users didn't provide a resource and there isn't a env var set. Use default one.
         temp_env::with_var_unset("OTEL_RESOURCE_ATTRIBUTES", || {
-            let default_config_provider = super::TracerProvider::builder().build();
+            let default_config_provider = super::SdkTracerProvider::builder().build();
             assert_resource(
                 &default_config_provider,
                 SERVICE_NAME,
@@ -603,7 +604,7 @@ mod tests {
         });
 
         // If user provided config, use that.
-        let custom_config_provider = super::TracerProvider::builder()
+        let custom_config_provider = super::SdkTracerProvider::builder()
             .with_resource(
                 Resource::builder_empty()
                     .with_service_name("test_service")
@@ -618,7 +619,7 @@ mod tests {
             "OTEL_RESOURCE_ATTRIBUTES",
             Some("key1=value1, k2, k3=value2"),
             || {
-                let env_resource_provider = super::TracerProvider::builder().build();
+                let env_resource_provider = super::SdkTracerProvider::builder().build();
                 assert_resource(
                     &env_resource_provider,
                     SERVICE_NAME,
@@ -636,7 +637,7 @@ mod tests {
             "OTEL_RESOURCE_ATTRIBUTES",
             Some("my-custom-key=env-val,k2=value2"),
             || {
-                let user_provided_resource_config_provider = super::TracerProvider::builder()
+                let user_provided_resource_config_provider = super::SdkTracerProvider::builder()
                     .with_resource(
                         Resource::builder()
                             .with_attributes([
@@ -678,7 +679,7 @@ mod tests {
         );
 
         // If user provided a resource, it takes priority during collision.
-        let no_service_name = super::TracerProvider::builder()
+        let no_service_name = super::SdkTracerProvider::builder()
             .with_resource(Resource::empty())
             .build();
 
@@ -689,7 +690,7 @@ mod tests {
     fn test_shutdown_noops() {
         let processor = TestSpanProcessor::new(false);
         let assert_handle = processor.assert_info();
-        let tracer_provider = super::TracerProvider::new(TracerProviderInner {
+        let tracer_provider = super::SdkTracerProvider::new(TracerProviderInner {
             processors: vec![Box::from(processor)],
             config: Default::default(),
             is_shutdown: AtomicBool::new(false),
@@ -704,7 +705,7 @@ mod tests {
 
         assert!(assert_handle.started_span_count(2));
 
-        let shutdown = |tracer_provider: super::TracerProvider| {
+        let shutdown = |tracer_provider: super::SdkTracerProvider| {
             let _ = tracer_provider.shutdown(); // shutdown once
         };
 
@@ -773,10 +774,10 @@ mod tests {
             });
 
             {
-                let tracer_provider1 = super::TracerProvider {
+                let tracer_provider1 = super::SdkTracerProvider {
                     inner: shared_inner.clone(),
                 };
-                let tracer_provider2 = super::TracerProvider {
+                let tracer_provider2 = super::SdkTracerProvider {
                     inner: shared_inner.clone(),
                 };
 
@@ -812,10 +813,10 @@ mod tests {
 
         // Create a scope to test behavior when providers are dropped
         {
-            let tracer_provider1 = super::TracerProvider {
+            let tracer_provider1 = super::SdkTracerProvider {
                 inner: shared_inner.clone(),
             };
-            let tracer_provider2 = super::TracerProvider {
+            let tracer_provider2 = super::SdkTracerProvider {
                 inner: shared_inner.clone(),
             };