+++ stress/src/traces.rs
@@ -11,7 +11,7 @@
 
 use lazy_static::lazy_static;
 use opentelemetry::{
-    trace::{Span, SpanBuilder, TraceResult, Tracer, TracerProvider as _},
+    trace::{Span, SpanBuilder, TraceResult, Tracer, TracerProvider},
     Context, KeyValue,
 };
 use opentelemetry_sdk::{
@@ -22,11 +22,11 @@ use opentelemetry_sdk::{
 mod throughput;
 
 lazy_static! {
-    static ref PROVIDER: sdktrace::TracerProvider = sdktrace::TracerProvider::builder()
+    static ref PROVIDER: sdktrace::SdkTracerProvider = sdktrace::SdkTracerProvider::builder()
         .with_sampler(sdktrace::Sampler::AlwaysOn)
         .with_span_processor(NoOpSpanProcessor {})
         .build();
-    static ref TRACER: sdktrace::Tracer = PROVIDER.tracer("stress");
+    static ref TRACER: sdktrace::SdkTracer = PROVIDER.tracer("stress");
 }
 
 #[derive(Debug)]